chore: fix PHP 8.5 ReflectionProperty deprecations and extract RegressionTestCase (#1181)
* chore: archive fix scripts from v2.25.6


* chore: fix PHP 8.5 ReflectionProperty deprecations and extract RegressionTestCase

- Remove ReflectionProperty::setAccessible() no-op calls from 7 test files
  (deprecated in PHP 8.5, has been a no-op since PHP 8.1)
- Collapse dead PHP_VERSION_ID >= 80100 guard in bootstrap-integration.php
  (project requires PHP 8.2+, condition was always true)
- Delete CarShowcaseServiceGetNewCarIdsTest — permanently skipped on any
  populated database via skipIfDatabaseIsPopulated(); existing
  CarShowcaseServiceTest and Playwright e2e coverage remain
- Extract RegressionTestCase base class from 4 regression test classes,
  eliminating 44 lines of duplicated $_POST/$_GET save/restore boilerplate;
  update RegressionTestTemplate to extend it
- Update pre-commit hook to allow non-Issue-prefixed RegressionTest files
  (RegressionTestCase, RegressionTestTemplate, EncodeAtOutputRegressionTest)

Closes #1151


* test: document parent::setUp() contract on RegressionTestCase

Adds inline comment explaining that subclasses overriding setUp() must
call parent::setUp() first — otherwise tearDown() restores [] instead
of the real pre-test $_POST/$_GET state.


---------
