+++ docs/releases/RELEASE_NOTES_v2.25.7.md
@@ -36,6 +36,6 @@
 - [#1096](https://github.com/unibrain1/elanregistry/issues/1096) — fix: correct DataTable catch behavior in car_details.js and resolve npm vulnerability
 - [#1126](https://github.com/unibrain1/elanregistry/issues/1126) — enhancement: add cache-busting version parameter to static asset URLs
 - [#1127](https://github.com/unibrain1/elanregistry/issues/1127) — feat: maintenance script — report and delete unverified accounts with no car associations
-- [#1151](https://github.com/unibrain1/elanregistry/issues/1151) — chore: fix PHP 8.5 ReflectionProperty deprecations in test infrastructure and remove unreliable CarShowcaseService tests
+- [#1151](https://github.com/unibrain1/elanregistry/issues/1151) ✓ — chore: fix PHP 8.5 ReflectionProperty deprecations in test infrastructure and remove unreliable CarShowcaseService tests
 - [#1167](https://github.com/unibrain1/elanregistry/issues/1167) — fix: CarRepository::getHistory() returns null for empty history — should return []
 - [#1175](https://github.com/unibrain1/elanregistry/issues/1175) — fix: process-transfer-approve.php — wrap transfer + updateStatus in a shared transaction
