+++ tests/bootstrap-integration.php
@@ -108,13 +108,7 @@ try {
         // The DB class caches the PDO connection, so we need to clear it to force a new one
         $reflectionClass = new ReflectionClass('DB');
         $instanceProperty = $reflectionClass->getProperty('_instance');
-        $instanceProperty->setAccessible(true);
-        // For static properties in PHP 8.1+, use static context
-        if (PHP_VERSION_ID >= 80100) {
-            // PHP 8.1+: Set static property directly via reflection
-            DB::class; // Ensure class is loaded
-            $instanceProperty->setValue(null);
-        }
+        $instanceProperty->setValue(null);
         fwrite(STDERR, "NOTE: Reset DB singleton cache for reinitialization\n");
 
         // Now create a fresh DB instance with the corrected configuration
