+++ tests/integration/ManageConsolidatedDeletionTest.php
@@ -40,9 +40,10 @@ final class ManageConsolidatedDeletionTest extends IntegrationTestCase
         $user = new User();
         $user->find(1);
 
+        // Bypass login() to set the private $_isLoggedIn flag directly via reflection.
+        // setAccessible() is intentionally omitted — it is a no-op since PHP 8.1.
         $reflection = new ReflectionClass($user);
         $isLoggedInProperty = $reflection->getProperty('_isLoggedIn');
-        $isLoggedInProperty->setAccessible(true);
         $isLoggedInProperty->setValue($user, true);
 
         $GLOBALS['user'] = $user;
